arg: move class
