+++ base/src/main/java/org/aya/core/ops/Eta.java
@@ -4,7 +4,7 @@ package org.aya.core.ops;
 
 import kala.collection.immutable.ImmutableSeq;
 import org.aya.core.term.*;
-import org.aya.generic.Arg;
+import org.aya.util.Arg;
 import org.aya.tyck.env.LocalCtx;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.VisibleForTesting;
