+++ base/src/main/java/org/aya/core/serde/Serializer.java
@@ -9,7 +9,7 @@ import kala.tuple.Tuple;
 import org.aya.core.def.*;
 import org.aya.core.pat.Pat;
 import org.aya.core.term.*;
-import org.aya.generic.Arg;
+import org.aya.util.Arg;
 import org.aya.generic.util.InternalException;
 import org.aya.ref.DefVar;
 import org.aya.ref.LocalVar;
