+++ base/src/main/java/org/aya/core/term/ListTerm.java
@@ -5,7 +5,7 @@ package org.aya.core.term;
 import kala.collection.immutable.ImmutableSeq;
 import org.aya.core.def.CtorDef;
 import org.aya.core.repr.AyaShape;
-import org.aya.generic.Arg;
+import org.aya.util.Arg;
 import org.aya.generic.Shaped;
 import org.jetbrains.annotations.NotNull;
