+++ base/src/main/java/org/aya/core/term/MetaTerm.java
@@ -5,7 +5,7 @@ package org.aya.core.term;
 import kala.collection.SeqView;
 import kala.collection.immutable.ImmutableSeq;
 import org.aya.core.Meta;
-import org.aya.generic.Arg;
+import org.aya.util.Arg;
 import org.jetbrains.annotations.NotNull;
 
 /**
