+++ base/src/main/java/org/aya/core/term/SigmaTerm.java
@@ -4,7 +4,7 @@ package org.aya.core.term;
 
 import kala.collection.immutable.ImmutableSeq;
 import org.aya.core.visitor.BetaExpander;
-import org.aya.generic.Arg;
+import org.aya.util.Arg;
 import org.aya.generic.SortKind;
 import org.aya.ref.LocalVar;
 import org.jetbrains.annotations.NotNull;
