+++ base/src/main/java/org/aya/tyck/pat/ClausesProblem.java
@@ -8,7 +8,7 @@ import org.aya.core.def.CtorDef;
 import org.aya.core.pat.Pat;
 import org.aya.core.term.DataCall;
 import org.aya.core.term.Term;
-import org.aya.generic.Arg;
+import org.aya.distill.BaseDistiller;
 import org.aya.pretty.doc.Doc;
 import org.aya.pretty.doc.Style;
 import org.aya.util.distill.DistillerOptions;
@@ -95,7 +95,7 @@ public sealed interface ClausesProblem extends Problem {
   ) implements ClausesProblem {
     @Override public @NotNull Doc describe(@NotNull DistillerOptions options) {
       return Doc.sep(Doc.english("Unhandled case:"),
-        Doc.commaList(pats.missing().map(t -> Arg.toDoc(options, t))));
+        Doc.commaList(pats.missing().map(t -> BaseDistiller.toDoc(options, t))));
     }
   }
